Refactor: Change hooks to contexts directory
