+++ src/components/navigation.js
@@ -1,7 +1,6 @@
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
-// import { withRouter } from "react-router";
 import { LinkContainer } from 'react-router-bootstrap'
 import '../assets/css/navigation.css';
